Move Arduino code to `src-old`
