+++ teensy/src-old/motor.cpp
